+++ OpenVBX/controllers/numbers.php
@@ -26,23 +26,38 @@ class Numbers extends User_Controller
 	private $error_message = FALSE;
 	private $new_number = null;
 
+	private $numbers_per_page = 50;
+
 	function __construct()
 	{
 		parent::__construct();
 		$this->section = 'numbers';
 		$this->template->write('title', 'Numbers');
 		$this->load->model('vbx_incoming_numbers');
+		$this->load->library('pagination');
 	}
 
 	function index()
 	{
 		$this->admin_only($this->section);
 		$this->template->add_js('assets/j/numbers.js');
-		
+
+		$max = $this->input->get_post('max');
+		$offset = $this->input->get_post('offset');
+
+		if (empty($offset)) {
+			$offset = 0;
+		}
+
+		if (empty($max)) {
+			$max = $this->numbers_per_page;
+		}
+
 		$data = $this->init_view_data();
 		$data['selected_country'] = $this->vbx_settings->get('numbers_country', $this->tenant->id);
 		
 		$numbers = array();
+		$total_numbers = 0;
 		$data['countries'] = array();
 		$data['openvbx_js']['countries'] = array();
 		try
@@ -85,10 +100,6 @@ class Numbers extends User_Controller
 				}
 				
 				$item->phone_formatted = format_phone($item->phone);
-				if (!empty($item->pin))
-				{
-					$item->pin = implode('-', str_split($item->pin, 4));
-				}
 
 				$capabilities = array();
 				if (!empty($item->capabilities)) 
@@ -103,7 +114,7 @@ class Numbers extends User_Controller
 				}
 				$item->capabilities = $capabilities;
 
-                $item->status = null;
+				$item->status = null;
 
 				if ($item->installed)
 				{
@@ -113,12 +124,12 @@ class Numbers extends User_Controller
 					
 					array_push($data['incoming_numbers'], $item);
 				}
-				elseif (!empty($item->url) || !empty($item->smsUrl))
+				elseif ((!empty($item->url) || !empty($item->smsUrl)) && $offset == 0)
 				{
 					// Number is in use elsewhere
 					array_push($data['other_numbers'], $item);
 				}
-				else
+				elseif ($offset == 0)
 				{
 					// Number is open for use
 					array_push($data['available_numbers'], $item);
@@ -148,6 +159,22 @@ class Numbers extends User_Controller
 		}
 
 		$data['counts'] = $this->message_counts();
+
+		/**
+		 * $numbers is a list of phone numbers straight from the Twilio API,
+		 * there's no fancy query logic here, just slicing up the array.
+		 */
+		$total_numbers = count($data['incoming_numbers']);
+		$data['incoming_numbers'] = array_slice($data['incoming_numbers'], $offset, $max, true);
+
+		// pagination
+		$page_config = array(
+			'base_url' => site_url('numbers'),
+			'total_rows' => $total_numbers,
+			'per_page' => $max
+		);
+		$this->pagination->initialize($page_config);
+		$data['pagination'] = CI_Template::literal($this->pagination->create_links());
 
 		$this->respond('', 'numbers/numbers', $data);
 	}